+++ test/support/erlang_style_codec.ex
@@ -9,26 +9,26 @@ defmodule ErlangStyleCodec do
   @type t :: String.t()
 
   @impl Spectral.Codec
-  def encode(_format, ErlangStyleCodec, {:type, :t, 0}, data, _sp_type, _params, _config)
+  def encode(_format, _caller_type_info, {:type, :t, 0}, _target_type, data, _config)
       when is_binary(data) do
     {:ok, data}
   end
 
-  def encode(_format, ErlangStyleCodec, {:type, :t, 0}, data, _sp_type, _params, _config) do
+  def encode(_format, _caller_type_info, {:type, :t, 0}, _target_type, data, _config) do
     {:error, [:sp_error.type_mismatch({:type, :t, 0}, data)]}
   end
 
-  def encode(_format, _module, _type_ref, _data, _sp_type, _params, _config), do: :continue
+  def encode(_format, _caller_type_info, _type_ref, _target_type, _data, _config), do: :continue
 
   @impl Spectral.Codec
-  def decode(_format, ErlangStyleCodec, {:type, :t, 0}, input, _sp_type, _params, _config)
+  def decode(_format, _caller_type_info, {:type, :t, 0}, _target_type, input, _config)
       when is_binary(input) do
     {:ok, input}
   end
 
-  def decode(_format, ErlangStyleCodec, {:type, :t, 0}, input, _sp_type, _params, _config) do
+  def decode(_format, _caller_type_info, {:type, :t, 0}, _target_type, input, _config) do
     {:error, [:sp_error.type_mismatch({:type, :t, 0}, input)]}
   end
 
-  def decode(_format, _module, _type_ref, _input, _sp_type, _params, _config), do: :continue
+  def decode(_format, _caller_type_info, _type_ref, _target_type, _input, _config), do: :continue
 end